trajectories visualisation

+++ src/main/java/entity/TrajectoryPoint.java
@@ -4,11 +4,11 @@ import java.util.Objects;
 
 public class TrajectoryPoint implements Cloneable {
 
-    private long x;
-    private long y;
+    private int x;
+    private int y;
     private long time;
 
-    public TrajectoryPoint(long x, long y) {
+    public TrajectoryPoint(int x, int y) {
         this.x = x;
         this.y = y;
     }
@@ -32,19 +32,19 @@ public class TrajectoryPoint implements Cloneable {
         return null;
     }
 
-    public long getX() {
+    public int getX() {
         return this.x;
     }
 
-    public void setX(long x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public long getY() {
+    public int getY() {
         return this.y;
     }
 
-    public void setY(long y) {
+    public void setY(int y) {
         this.y = y;
     }
